fix: change the number of blocks
I changed the number of blocks 32768 to 4096. Because NAND packages are
accessed by chip based NOT package based. For this reason, each
package's chip has (number of blocks in a NAND package / number of chips
in a NAND package).

However, current flash board might be not supported to LUN selection.
For this reason, we can only use the half size of a chip. In other
words, we can only access the 4096 blocks NOT 8192 blocks.